requirements.txt
# FastAPI
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# Email
aiosmtplib>=3.0.0

# Template
jinja2>=3.1.0

# MCP
mcp[cli]>=1.0.0
httpx>=0.27.0

# Scheduler
apscheduler>=3.10.0

# Config
pyyaml>=6.0

# IMAP (bounce detection)
aioimaplib>=1.1.0

# HTML parsing (link rewrite)
beautifulsoup4>=4.12.0

# Encryption
cryptography>=42.0.0
